move jpa entity to persistence package
